refactor: extract prompt building (stage 1,2) into separate file
- Move stage 1 (prompt building) and stage 2 (memory processing) logic
  to index_promptBuild.svelte.ts
- Add buildPrompt function with typed input/output interface
- Simplify index.svelte.ts to focus on stage 3 (AI request) and
  stage 4 (post processing)
